A bunch of refactoring
Why:

*

Clean up file strcuture, change setStates to use
function.

*
